fix(decorators): separate class, create local func
