+++ src/Serilog.Ui.Web/Endpoints/SerilogUiAppRoutesDecorator.cs
@@ -20,20 +20,16 @@ namespace Serilog.Ui.Web.Endpoints
 
         public Task GetHome(HttpContext httpContext)
         {
-#if NET6_0_OR_GREATER
-            Guard.Against.Null(Options);
-#else
             Guard.Against.Null(Options, nameof(Options));
-#endif
 
-            var changeResponse = (HttpResponse response) =>
+            static Task ChangeResponseAsync(HttpResponse response)
             {
                 response.ContentType = "text/html;charset=utf-8";
                 return response.WriteAsync("<p>You don't have enough permission to access this page!</p>", Encoding.UTF8);
-            };
+            }
 
             return Options.Authorization.RunAuthorizationFilterOnAppRoutes ?
-                _authFilterService.CheckAccess(httpContext, Options, _decoratedService.GetHome, changeResponse) :
+                _authFilterService.CheckAccess(httpContext, Options, _decoratedService.GetHome, ChangeResponseAsync) :
                 _decoratedService.GetHome(httpContext);
         }
 
@@ -48,36 +44,6 @@ namespace Serilog.Ui.Web.Endpoints
             return Options.Authorization.RunAuthorizationFilterOnAppRoutes ?
                 _authFilterService.CheckAccess(httpContext, Options, _decoratedService.RedirectHome) :
                 _decoratedService.RedirectHome(httpContext);
-        }
-
-        public void SetOptions(UiOptions options)
-        {
-            Options = options;
-            _decoratedService.SetOptions(options);
-        }
-    }
-
-    internal class SerilogUiEndpointsDecorator : ISerilogUiEndpoints
-    {
-        private readonly ISerilogUiEndpoints _decoratedService;
-        private readonly IAuthorizationFilterService _authFilterService;
-
-        public SerilogUiEndpointsDecorator(ISerilogUiEndpoints decoratedService, IAuthorizationFilterService authFilterService)
-        {
-            _decoratedService = decoratedService;
-            _authFilterService = authFilterService;
-        }
-
-        public UiOptions Options { get; private set; }
-
-        public Task GetApiKeys(HttpContext httpContext)
-        {
-            return _authFilterService.CheckAccess(httpContext, Options, _decoratedService.GetApiKeys);
-        }
-
-        public Task GetLogs(HttpContext httpContext)
-        {
-            return _authFilterService.CheckAccess(httpContext, Options, _decoratedService.GetLogs);
         }
 
         public void SetOptions(UiOptions options)
